Enable multiple package deployments
ServiceRunner now supports updating existing deployed projects. This
change removes the check in the CLI preventing the user from deploying
if there was an existing deployment. Additionally, for packagse which
have reached production this change also providers the user with an
example curl command for invoking their API from the shell.