Warn rather than fail if container is missing a supported language
Currently the bopmatic CLI will early exit if it fails to obtain a
directory listing from the bopmatic build container for a given
supported language. However, if a customer has an older version of the
bopmatic build container it may not contain examples for a given
supported language. With this change the CLI will instead just warn
the user and skip templates from that language rather than fail.