Update to supported refactored sdk
The recent SR API refactor series of changes to the sdk separates the
notions of deployments from packages and introduces project
registration. This commit is the first in a series to update the CLI
accordingly. This commit introduces new 'project' and 'deploy'
commands and moves the older 'new' command into a 'project create'
subcommand.